backend/lambda-functions/ai-handler/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
PyPDF2>=3.0.0
python-docx>=0.8.11
